nugenomics-project/AI-project-mainold/nugenomics-project/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.8.3
google-adk==0.3.0
litellm==1.35.8
requests==2.31.0
beautifulsoup4==4.12.3
aiohttp==3.9.5
rich==13.7.0
pandas==2.2.3
